client/requirements.txt
hazelcast-python-client[stats]==5.3.0
pyyaml==6.0.1